text along path
The attached diff implements typesetting of text along
a path as a decorator (curvedtext). I put it in the bug
tracker so that it doesn't get lost. It's a diff to
r2824 (taking into account the reversal of trafos in
canvas.py (r2728->r2730)).

From the docstring:
    """a text decorator for curved text

    - text: is typeset along the path to which this
decorator is applied
    - relarclenpos: position for the base point of the
text (default: 0)
    - arlenfrombegin, arclenfromend: alternative ways
of specifying the position of the base point;
                                     use of
relarclenpos, arclenfrombegin and arclenfromend is
mutually exclusive
    - textattrs, texrunner: standard text arguments
(defaults: [] resp None)
    """

Files touched aside from deco.py:

dvifile.__init__() in dvifile.py gets a new parameter
singlecharmode which defaults to 0; dvifile.putchar()
switches to single character mode if set to 1.

texrunner.__init__(), texrunner.finishdvi(),
texrunner.text() and texrunner.textboxes() in text.py
get the same parameter so that it can be passed from
the texrunner to dvifile.

curvedtext switches to singlecharmode only while it is
needed.

Basic idea for curvedtext goes back to schindmi, I think.