graph.style.surface: Make mincolor and maxcolor work
Currently, the mincolor and maxcolor arguments to surface are plainly
ignored.

Make them work as described, i.e. overriding the default coming from the
actual range of values of the color column. In particular, this makes it
possible to draw a surface with a single color without raising a
division by zero.

Also, make sure that actual color values are cut off at the forced min
and max.

Signed-off-by: Michael J Gruber <michaeljgruber@users.sourceforge.net>