deco.arrow(): allow relative positioning of the arrow
Currently, the pos parameter gives the position of the head of tip of
the arrow. This is problematic when one wants to position the arrow in
the middle of a line or arc: pos=0.5 gives an arrow which is visually
off the middle (it is visually to the left of it).

Introduce a parameter poscorr which shifts the arrow by an amount which
is relative to the size of the arrowhead (and the constriction).

poscorr=0 is the default (tip at pos)
poscorr=1 puts the constriction center at pos