Fix epsfile(kpsearch=1)
b1b3a96 (use new filelocator code, 2011-05-14) broke kpsearch
functionality for epsfile.epsfile() because use of the filelocator was
moved after the determination of an initial bbox, which has to open the
file unless a bbox was specified.

Make _readbox() pass kpsearch and appropriate formats to
linefilereader(), and make linefilereader() use the new filelocator
accordingly.

Signed-off-by: Michael J Gruber <michaeljgruber@users.sourceforge.net>