arrow: make pos relative to the full arclength
Currently, pos is relative to (arclength - constrictionlen) which makes
it very hard to position arrow heads correctly at, say, quarter circles etc.

Make it relative to the full arclength to ease positioning of the
arrow heads. This requires putting in a safeguard (so that arrow heads
do no come out before nor after that path) which may be of independent
interest.