deco.arrow: fix arrows at pos=0
The arrowhead needs a piece of size "size" of the path for its
construction. Therefore, only (arclen-size) of the path is available as
positions for a head without extending the path.

So, change pos=0 to mean a position at arclenpos=size, and introduce the
necessary additional path exclusions for begin/end arrows.