do not shift grids when shifting axes
"?axisat" allow shifting the axes to other locations, e.g. the quite
common case of having the axes go through 0. This is implemented by
shifting all axis layers, which includes the grid layer, so that it gets
placed outside of the graph.

Omit the grid layer when shifting around axis layers so that the grid
remains in place.