allow Lipshitz bounds on function plots
Depending on the number of points, PyX may or may not catch
singularities such as that of 1/x at 0. This may lead to spurious
vertical lines.

Introduce an extra parameter "lipshitz" to graph.data.function(). It
defaults to off. When specified and when two subsequent graph points
produce a slope larger in abosulte value than "lipshitz", an extra
"None" value is aded to the points produced by graph.data.function().
This makes graph.style.line() cut the plot between these two points.

This also catches discontinuities.